email_otp_server/requirements.txt
fastmcp>=0.5.0
uvicorn==0.38.0
fastapi==0.115.5
python-dotenv==1.0.1
azure-communication-email==1.0.0
pydantic==2.12.3